@@ -16,50 +16,54 @@
   "author": "thomas caillard",
   "license": "ISC",
   "devDependencies": {
-    "@types/lodash": "4.14.123",
-    "@types/react": "16.8.13",
-    "@types/react-dom": "16.8.4",
-    "@types/reactstrap": "8.0.0",
+    "@testing-library/react": "^9.4.0",
+    "@types/jest": "^25.1.2",
+    "@types/lodash": "4.14.149",
+    "@types/react": "16.9.19",
+    "@types/react-dom": "16.9.5",
+    "@types/reactstrap": "8.4.1",
     "@types/rx": "^4.1.1",
     "bootstrap": "^4.1.1",
-    "clean-webpack-plugin": "^1.0.1",
-    "css-loader": "^2.1.0",
+    "clean-webpack-plugin": "^3.0.0",
+    "css-loader": "^3.4.2",
     "html-webpack-plugin": "^3.2.0",
     "jquery": "^3.3.1",
-    "json-server": "^0.14.0",
+    "json-server": "^0.15.1",
     "lodash": "^4.17.10",
     "prettier": "^1.16.1",
     "prop-types": "^15.6.2",
     "react": "^16.8.6",
     "react-dom": "^16.8.6",
-    "reactstrap": "8.0.0",
+    "reactstrap": "8.4.1",
     "redux-observable": "^1.0.0",
-    "rxjs": "6.4.0",
-    "ts-loader": "^5.3.3",
-    "typescript": "3.4.3",
+    "rxjs": "6.5.4",
+    "ts-jest": "^25.2.0",
+    "ts-loader": "^6.2.1",
+    "typescript": "3.7.5",
     "typescript-fsa": "^3.0.0-beta-2",
-    "webpack": "4.30.0",
-    "webpack-cli": "3.3.0",
-    "webpack-dev-server": "3.3.1"
+    "webpack": "4.41.5",
+    "webpack-cli": "3.3.10",
+    "webpack-dev-server": "3.10.3"
   },
   "dependencies": {
-    "@types/react-redux": "^7.0.0",
-    "@types/react-router": "^4.0.28",
+    "@types/react-redux": "^7.1.7",
+    "@types/react-router": "^5.1.4",
     "@types/react-router-bootstrap": "^0.24.5",
-    "@types/react-router-dom": "^4.2.7",
+    "@types/react-router-dom": "^5.1.3",
     "@types/redux-logger": "^3.0.6",
-    "axios": "^0.18.0",
+    "axios": "^0.19.2",
     "font-awesome": "^4.7.0",
+    "jest": "^25.1.0",
     "popper.js": "^1.14.3",
-    "react-redux": "^6.0.0",
-    "react-router": "^4.3.1",
-    "react-router-bootstrap": "^0.24.4",
-    "react-router-dom": "^4.3.1",
-    "react-toastify": "^4.1.0",
-    "reduce-reducers": "1.0.1",
-    "redux": "^4.0.0",
+    "react-redux": "^7.1.3",
+    "react-router": "^5.1.2",
+    "react-router-bootstrap": "^0.25.0",
+    "react-router-dom": "^5.1.2",
+    "react-toastify": "^5.5.0",
+    "reduce-reducers": "1.0.4",
+    "redux": "^4.0.5",
     "redux-logger": "^3.0.6",
     "redux-saga": "^1.0.0",
-    "style-loader": "^0.23.1"
+    "style-loader": "^1.1.3"
   }
 }
