@@ -1,23 +1,11 @@
 import { mapperToPlayerInfo, mapperToPlayerChest } from "./mapper";
-import { PlayerInfoDto } from "./typings";
 import { PlayerInfo, PlayerChest } from "../../components/PlayerClash.typings";
+import instance from "../../axios";
 
-import axiosInstance from "../../axios";
-import { from, Observable } from "rxjs";
-import { map } from "rxjs/operators";
-
-// export const postComment = (body: string, postId: number) =>
-//   fetch(`${baseUrl}/comments/`, {
-//     method: "post",
-//     headers: {
-//       Accept: "application/json, text/plain, */*",
-//       "Content-Type": "application/json"
-//     },
-//     body: JSON.stringify({ body, postId })
-//   }).then(res => res.json());
+const baseUrl = "/api";
 
 const fetchByTypeById = <T>(type: string, id: string): Promise<T> => {
-  const fetchCall = axiosInstance.get(`/${type}/${id}`);
+  const fetchCall = instance.get(`${baseUrl}/${type}/${id}`);
   return fetchCall
     .then(response => response.data)
     .catch(e => {
@@ -26,19 +14,13 @@ const fetchByTypeById = <T>(type: string, id: string): Promise<T> => {
 };
 
 const fetch = <T>(url: string): Promise<T> => {
-  const fetchCall = axiosInstance.get(`/${url}`);
+  const fetchCall = instance.get(`${baseUrl}/${url}`);
   return fetchCall
     .then(response => response.data)
     .catch(e => {
       throw e;
     });
 };
-
-// export const fetchPlayerInfoObservable = (
-//   type: string,
-//   id: string
-// ): Observable<PlayerInfo> =>
-//   from<PlayerInfoDto>(fetchByTypeById(type, id)).pipe(map(mapperToPlayerInfo));
 
 export const fetchPlayerInfo = (
   type: string,
@@ -49,4 +31,4 @@ export const fetchPlayerChest = (
   type: string,
   id: string
 ): Promise<PlayerChest> =>
-  fetch(`${type}/${id}/chest`).then(mapperToPlayerChest);
+  fetch(`${type}/${id}/chests`).then(mapperToPlayerChest);
